+++ src/main/resources/createusertable.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
     player_id BINARY(16) NOT NULL,
-    coins INT NOT NULL DEFAULT 0,
+    coins BIGINT NOT NULL DEFAULT 0,
     PRIMARY KEY (player_id)
 );
